requirements.txt
aiogram>=3.4.1
apscheduler>=3.10.4
pandas>=2.0.0
python-dotenv>=1.0.1
aiosqlite>=0.20.0
aiohttp>=3.9.0
numpy<2.0.0

